Changed double to float in books.py
Also removed some tabs

+++ src/web/books.py
@@ -12,7 +12,7 @@ def render_create_book(request):
         response = HttpResponse()
         response.write(loader.render_to_string(tmpl, context))
         return response
-	
+
 def create_book_action(request):
     isbn = cgi.escape(request.POST['isbn'])
     title = cgi.escape(request.POST['title'])
@@ -20,17 +20,17 @@ def create_book_action(request):
     year = int(cgi.escape(request.POST['year']))
     edition = cgi.escape(request.POST['edition'])
     publisher = cgi.escape(request.POST['publisher'])
-    rrp = double(cgi.escape(request.POST['rrp']))
+    rrp = float(cgi.escape(request.POST['rrp']))
     picture = cgi.escape(request.POST['picture'])
     rrp = int(rrp * 100) #convert P.pp to interger pence
-	
+
     context = Context()
     try:
         BOOK.create_book(isbn, title, author, year, edition, publisher, rrp, picture)
         tmpl =  os.path.join(os.path.dirname(__file__), 'template', 'create_book_success.html')
     except:
         tmpl =  os.path.join(os.path.dirname(__file__), 'template', 'create_book_failure.html')
-		
+
     response = HttpResponse()
     response.write(loader.render_to_string(tmpl, context))
     return response
